+++ src/views/Settings/SettingsResetPassword.vue
@@ -64,6 +64,7 @@ import { touchKeystore, initWallet } from '@/actions/vue/create-wallet'
 import FormErrorMessage from '@/components/FormErrorMessage.vue'
 import FormField from '@/components/FormField.vue'
 import ButtonSubmit from '@/components/ButtonSubmit.vue'
+import { radixConnection } from '@/helpers/network'
 
 interface PasswordForm {
   currentPassword: string;
@@ -85,15 +86,13 @@ const SettingsResetPassword = defineComponent({
     const isLoading = ref(false)
     const updatedPassword: Ref<boolean> = ref(false)
 
-    const radix = Radix
-      .create({ network: Network.STOKENET })
-      .connect(process.env.VUE_APP_API || 'https://stokenet.radixdlt.com')
+    const radix = radixConnection()
       .withWallet(store.state.wallet)
 
     const handleResetPassword = (newPassword: string) => {
       const getMnemonicForPassword = radix.revealMnemonic()
         .subscribe((m: MnemomicT) => {
-          initWallet(m, newPassword, Network.STOKENET) // Temporarily hardcoded for betanet
+          initWallet(m, newPassword, Network.MAINNET)
             .then(wallet => {
               store.commit('setWallet', wallet)
               resetForm()